+++ OptimizelySDKUniversal/generated-frameworks/Release-iOS-universal-SDK/OptimizelySDKiOS.framework/Headers/OPTLYEventBuilder.h
@@ -43,7 +43,7 @@ NS_ASSUME_NONNULL_END
 - (nullable NSDictionary *)buildImpressionEventForUser:(nonnull NSString *)userId
                                             experiment:(nonnull OPTLYExperiment *)experiment
                                              variation:(nonnull OPTLYVariation *)variation
-                                            attributes:(nullable NSDictionary<NSString *, NSObject *> *)attributes;
+                                            attributes:(nullable NSDictionary<NSString *, id> *)attributes;
    
 /**
  * Create the parameters for a conversion event.
@@ -58,7 +58,7 @@ NS_ASSUME_NONNULL_END
 - (nullable NSDictionary *)buildConversionEventForUser:(nonnull NSString *)userId
                                                  event:(nonnull OPTLYEvent *)event
                                              eventTags:(nullable NSDictionary *)eventTags
-                                            attributes:(nullable NSDictionary<NSString *, NSObject *> *)attributes;
+                                            attributes:(nullable NSDictionary<NSString *, id> *)attributes;
 @end
 
 @interface OPTLYEventBuilderDefault : NSObject<OPTLYEventBuilder>